+++ tests/integration/cli.edge-cases.test.js
@@ -45,9 +45,8 @@ describe('integration.cli.edge-cases', () => {
     const logFileBase = join(tmpRoot, 'unsupported-npm-version')
     const cwd = join(dummyProjectsRoot, 'with-lockfile')
 
-    // lowest supported = [9.0.0] - need to find a lower number
     const npmVersion = [
-      Math.round(NPM_LOWEST_SUPPORTED[0] * Math.random()),
+      Math.round((NPM_LOWEST_SUPPORTED[0] - 1) * Math.random()),
       Math.round(99 * Math.random()),
       Math.round(99 * Math.random())
     ]