+++ HISTORY.md
@@ -8,9 +8,10 @@ All notable changes to this project will be documented in this file.
 
 
 * BREAKING Changes
-  * Dropped support for `node<20.18.0` ([#1192] via [#])
+  * Dropped support for `node<20.18.0` ([#1192] via [#1273])
 
 [#1192]: https://github.com/CycloneDX/cyclonedx-node-npm/issues/1192
+[#1273]: https://github.com/CycloneDX/cyclonedx-node-npm/pull/1273
 
 ## 2.1.0 - 2025-03-19