+++ HISTORY.md
@@ -10,6 +10,10 @@ All notable changes to this project will be documented in this file.
 * BREAKING Changes
   * Dropped support for `node<20.18.0` ([#1192] via [#1273])
   * Dropped support for `npm<9` ([#1274] via [#1273], [#1277])
+* Runtime Dependencies
+    * Raised `@cyclonedx/cyclonedx-library@^8.0.0`, was `@^7.0.0` (via [#1281])
+    * Raised `commander@^13.1.0`, was `@^10.0.0` (via [#1281])
+    * Raised `normalize-package-data@^7.0.0`, was `@^3||^4||^5||^6` (via [#1281])
 * Build
   * Use _TypeScript_ `v5.8.2` now, was `v5.7.3` (via [#1267])
 
@@ -18,6 +22,7 @@ All notable changes to this project will be documented in this file.
 [#1273]: https://github.com/CycloneDX/cyclonedx-node-npm/pull/1273
 [#1274]: https://github.com/CycloneDX/cyclonedx-node-npm/issues/1274
 [#1277]: https://github.com/CycloneDX/cyclonedx-node-npm/pull/1277
+[#1281]: https://github.com/CycloneDX/cyclonedx-node-npm/pull/1281
 
 ## 2.1.0 - 2025-03-19